Deploys homework application on Kubernetes
Deploys an application with a Deployment, Namespace, and ConfigMap.

The Deployment defines the application's replicas, container image,
resource requirements, and node affinity. It utilizes an init container
to fetch static content.  It also configures a readiness probe and preStop hook.

The Namespace isolates the application resources.

The ConfigMap provides configuration data for the application.